+++ src/Checkout/ExpressCheckout/Service/PayPalExpressCheckoutDataService.php
@@ -72,8 +72,8 @@ class PayPalExpressCheckoutDataService
             'currency' => $salesChannelContext->getCurrency()->getIsoCode(),
             'intent' => $settings->getIntent(),
             'addProductToCart' => $addProductToCart,
-            'createPaymentUrl' => $this->router->generate('sales-channel-api.action.paypal.create_payment', ['version' => 1]),
-            'createNewCartUrl' => $this->router->generate('sales-channel-api.action.paypal.create_new_cart', ['version' => 1]),
+            'createPaymentUrl' => $this->router->generate('sales-channel-api.action.paypal.create_payment', ['version' => 2]),
+            'createNewCartUrl' => $this->router->generate('sales-channel-api.action.paypal.create_new_cart', ['version' => 2]),
             'addLineItemUrl' => $this->router->generate('frontend.checkout.line-item.add'),
             'approvePaymentUrl' => $this->router->generate('paypal.approve_payment'),
             'checkoutConfirmUrl' => $this->router->generate(